Borrowing package done. Dates don't save yet.

+++ src/dao/custom/BorrowingsDao.java
@@ -3,6 +3,6 @@ package dao.custom;
 import dao.CrudDao;
 import entity.Borrowings;
 
-public interface BorrowingsDao extends CrudDao<Borrowings, Integer> {
+public interface BorrowingsDao extends CrudDao<Borrowings, Long> {
     
 }
